control parameters

control_calib is function that contains control parameters for joint_calib_create_matrix

Usage

control_calib(
  interpolation = c("logit", "linear"),
  logit_const = -1000,
  sum_to_sample = FALSE,
  sum_to_one = FALSE,
  survey_sparse = FALSE,
  ebal_constraint_tolerance = 1,
  ebal_print_level = 0,
  el_att = FALSE
)

Arguments

interpolation

type of interpolation: logit or linear,

logit_const

constant for logit interpolation,

sum_to_sample

whether weights should sum to sample,

sum_to_one

whether weights should sum to one (aka normalized weights),

survey_sparse

whether to use sparse matrices via Matrix package in survey::grake() (currently not supported),

ebal_constraint_tolerance

this is the tolerance level used by ebalance to decide if the moments in the reweighted data are equal to the target moments (see ebal::ebalance()),

ebal_print_level

controls the level of printing: 0 (normal printing), 2 (detailed), and 3 (very detailed) (see ebal::ebalance()),

el_att

whether weights for control should sum up to treatment size (for calib_el function only).

Value

a list with parameters

Author

Maciej Beręsewicz